+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+# domino x.x.x (not yet released)
+
 # domino 1.0.26 (15 Oct 2016)
 * Implement Document#dir.
 * Minor spec-compliance fixes to Document#title and classList#contains.