+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # domino x.x.x (not yet released)
+* Fix bug in AFE list replacement over existing bookmark.
 
 # domino 1.0.26 (15 Oct 2016)
 * Implement Document#dir.