+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Fix bug in AFE list replacement over existing bookmark.
 * Update htmlwg test suite to latest w3c/web-platform-tests.
 * Update html5lib test suite to latest.
+* HTML5 spec update: <menuitem> is no longer an empty element.
 
 # domino 1.0.26 (15 Oct 2016)
 * Implement Document#dir.