+++ CHANGELOG.md
@@ -11,6 +11,7 @@
   `rb`, `rtc`, `strike`, and `tt`.
 * HTML5 spec update: Remove extra newline in serialization of `<pre>`,
   `<listing>`, `<textarea>`. (#88)
+* HTML5 spec update: Remove case normalization for defunct SVG attributes.
 
 # domino 1.0.26 (15 Oct 2016)
 * Implement Document#dir.