+++ CHANGELOG.md
@@ -13,6 +13,7 @@
   `<listing>`, `<textarea>`. (#88)
 * HTML5 spec update: Remove case normalization for defunct SVG attributes.
 * Implement HTMLMenuItemElement#label.
+* Basic SVG support. (#81, #82)
 
 # domino 1.0.26 (15 Oct 2016)
 * Implement Document#dir.