+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# domino x.x.x (not yet released)
+# domino 1.0.27 (17 Oct 2016)
 * Fix bug in AFE list replacement over existing bookmark.
 * Update htmlwg test suite to latest w3c/web-platform-tests.
 * Update html5lib test suite to latest.